+++ packages/node/src/clients/renegade/admin.ts
@@ -1,0 +1,79 @@
+import { createConfig } from "@renegade-fi/core";
+import {
+    type CreateOrderInMatchingPoolParameters,
+    createOrderInMatchingPool,
+} from "@renegade-fi/core/actions";
+import * as rustUtils from "../../../renegade-utils/index.js";
+import { RenegadeClient } from "./base.js";
+import type { ConstructorParams } from "./types.js";
+
+export class AdminRenegadeClient extends RenegadeClient {
+    private readonly apiKey: string;
+
+    /**
+     * @internal
+     */
+    private constructor(params: ConstructorParams & { apiKey: string }) {
+        const { apiKey, ...renegadeParams } = params;
+        super(renegadeParams);
+        this.apiKey = apiKey;
+    }
+
+    /**
+     * Create an admin client for any chain by seed.
+     *
+     * @param params.chainId – the chain ID (e.g. CHAIN_IDS.ArbitrumMainnet)
+     * @param params.seed    – your 0x… seed
+     */
+    static override new({
+        chainId,
+        seed,
+        apiKey,
+    }: {
+        chainId: number;
+        seed: `0x${string}`;
+        apiKey: string;
+    }): AdminRenegadeClient {
+        return new AdminRenegadeClient({ chainId, mode: "seed", seed, apiKey });
+    }
+
+    /**
+     * @see AdminRelayerClient.new
+     */
+    static override newWithExternalKeychain: never;
+    /**
+     * @see AdminRelayerClient.new
+     */
+    static override newArbMainnetClient: never;
+    /**
+     * @see AdminRelayerClient.new
+     */
+    static override newArbMainnetClientWithKeychain: never;
+    /**
+     * @see AdminRelayerClient.new
+     */
+    static override newArbSepoliaClient: never;
+    /**
+     * @see AdminRelayerClient.new
+     */
+    static override newArbSepoliaClientWithKeychain: never;
+
+    /**
+     * Create an order in a matching pool
+     *
+     * @param params.matchingPool – the matching pool address
+     */
+    async createOrderInMatchingPool(params: CreateOrderInMatchingPoolParameters) {
+        const config = createConfig({
+            darkPoolAddress: this.configv2.darkpoolAddress,
+            priceReporterUrl: this.configv2.priceReporterUrl,
+            relayerUrl: this.configv2.relayerUrl,
+            chainId: this.configv2.id,
+            utils: rustUtils,
+            // Inject the admin key
+            adminKey: this.apiKey,
+        });
+        config.setState((s) => ({ ...s, seed: this.seed }));
+        return createOrderInMatchingPool(config, params);
+    }
+}
